Hoops
Hoops
Hoops
Hoooooops
In case I needed more to jump through.
Submitted the Garmin app for review
Submitted the Android app
And this is just all the pretty pictures I had to make. And a couple minor tweaks to the code bc I always forget to updated the version number